Backend/requirements.txt
requests~=2.27.1
PyYAML~=5.4.1
beautifulsoup4~=4.10.0
statham~=0.12
soupsieve~=2.3.1
pip~=20.2.3
MarkupSafe~=2.1.0
Werkzeug~=2.0.3
Jinja2~=3.0.3
click~=8.0.4
itsdangerous~=2.1.0
certifi~=2021.10.8
python-dateutil~=2.8.2
six~=1.16.0
idna~=3.3
urllib3~=1.26.8
setuptools~=49.2.1
jsonpointer~=2.2
Flask~=2.0.3
lxml=~4.8.0